create the class HomeActivity
the logic for the class homeActivity has been created
